Project Finished

+++ frontend/src/components/project/discussion/DiscussionRow.js
@@ -1,22 +1,31 @@
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 import { connect } from 'react-redux';
 
-import {useHttpClient} from "../../../hooks/http-hook";
-import {deleteDiscussion} from "../../../actions/project-action";
+import { useHttpClient } from "../../../hooks/http-hook";
+import { deleteDiscussion } from "../../../actions/project-action";
 import './DiscussionRow.css';
 
 const DiscussionRow = ({ discussion, username, handleClickOnEdit, projectId, deleteDiscussion }) => {
     const { sendRequest } = useHttpClient();
+    const [ isMobile, setIsMobile ] = useState(false);
 
-    const handleEditClick = () => {
-        handleClickOnEdit(discussion._id, discussion.text);
+    const handleEditClick = async () => {
+        await handleClickOnEdit(discussion._id, discussion.text);
     }
 
     const handleDeleteClick = async () => {
-        await deleteDiscussion(projectId, discussion._id, sendRequest);
+        if(window.confirm('Do you want to delete this discussion?')) {
+            await deleteDiscussion(projectId, discussion._id, sendRequest);
+        }
     }
+
+    useEffect(() => {
+        if (/Mobi/.test(navigator.userAgent))
+            setIsMobile(true);
+    }, [])
+
     return (
-        <div className="row white discussion-row showEditDeleteOnHover">
+        <div className={`row white discussion-row ${isMobile ? '' : 'showEditDeleteOnHover'}`} id="discussion-row">
             <div className="col s1 discussion-row__image">
                 <img
                     src={discussion?.user?.profileImage?.imageUrl}
@@ -30,8 +39,8 @@ const DiscussionRow = ({ discussion, username, handleClickOnEdit, projectId, del
                     <a href={`/member/${discussion.user.username}`}> {discussion.user.username}</a>
                     {username && (username === discussion.user.username) && (
                         <>
-                            <span className="edit" onClick={handleEditClick}>Edit</span>
-                            <span className="delete" onClick={handleDeleteClick}>Delete</span>
+                            <span id='edit' className={`edit ${isMobile ? 'showEdit' : ''}`} onClick={handleEditClick}>Edit</span>
+                            <span id='delete' className={`delete ${isMobile ? 'showDelete' : ''}`} onClick={handleDeleteClick}>Delete</span>
                         </>
                     )}
                 </p>
